requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==1.10.15
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
httpx==0.27.0
